improve runtime of diff unit tests
